HPCC-13909 Deprecated warnings from ldap code
The following deprecated functions are still being called by our LDAP code
and need to be updated

 CURRENT               NEWEST
 ldap_init		    -> ldap_initialize
 ldap_simple_bind_s	-> ldap_sasl_bind_s
 ldap_unbind		    -> ldap_unbind_ext
 ldap_get_values	   -> ldap_get_values_len
 ldap_value_free	   -> ldap_value_free_len
 ldap_compare_s	    -> ldap_compare_ext_s
 ldap_modify_s	     -> ldap_modify_ext_s
 ldap_delete_s	     -> ldap_delete_ext_s
 ldap_search_s	     -> ldap_search_ext_s

This PR makes the above changes, with the exception of the simple bind, which
is a change to the way we authenticate and is out of the scope of this bug.
However, I did move those calls to a new "ldaputils" source file so they
warnings would not appear when the headers were included
