+++ system/security/LdapSecurity/ldapconnection.hpp
@@ -40,6 +40,14 @@
 #endif
 
 #ifdef _WIN32
+    #define LDAP_COMPARE_EXT_S(ld,dn,attr,bval,svrctrls,clientctrls,msgnum) ldap_compare_ext_s(ld,(char*)dn,(char*)attr,(struct berval *)bval,svrctrls,clientctrls,msgnum)
+    #define LDAP_UNBIND(ld) ldap_unbind(ld)
+#else
+    #define LDAP_COMPARE_EXT_S(ld,dn,attr,bval,svrctrls,clientctrls,msgnum) ldap_compare_ext_s(ld,(char*)dn,(char*)attr,(struct berval *)bval,svrctrls,clientctrls)
+    #define LDAP_UNBIND(ld) ldap_unbind_ext(ld,0,0)
+#endif
+
+#ifdef _WIN32
     typedef struct l_timeval TIMEVAL;
 #else
     typedef struct timeval TIMEVAL;
@@ -199,293 +207,48 @@ ILdapClient* createLdapClient(IPropertyTree* cfg);
 bool verifyServerCert(LDAP* ld, PCCERT_CONTEXT pServerCert);
 #endif
 
-class LdapUtils
+
+//--------------------------------------------
+// This helper class ensures memory allocated by
+// calls to ldap_get_values_len gets freed
+//--------------------------------------------
+class CLDAPGetValuesLenWrapper
 {
+private:
+    struct berval** bvalues;
+    unsigned numValues;
 public:
-    static LDAP* LdapInit(const char* protocol, const char* host, int port, int secure_port)
+    CLDAPGetValuesLenWrapper()
     {
-        LDAP* ld = NULL;
-        if(stricmp(protocol, "ldaps") == 0)
-        {
+        bvalues = NULL;
+        numValues = 0;
+    }
+    CLDAPGetValuesLenWrapper(LDAP *ld, LDAPMessage *msg, const char * attr)
+    {
+        bvalues = NULL;
+        retrieveBValues(ld,msg,attr);
+    }
+
+    ~CLDAPGetValuesLenWrapper()
+    {
+        if (bvalues)
+            ldap_value_free_len(bvalues);
+    }
+    inline bool hasValues()         { return bvalues != NULL  && *bvalues != NULL; }
+    inline berval **queryBValues()  { return bvalues; }
+    inline const char * queryCharValue(unsigned which){ return which < numValues ? (*(bvalues[which])).bv_val : NULL; }
+
+    //Delayed call to ldap_get_values_len
+    void retrieveBValues(LDAP *ld, LDAPMessage *msg, const char * attr)
+    {
+        if (bvalues)
+            ldap_value_free_len(bvalues);
 #ifdef _WIN32
-            ld = ldap_sslinit((char*)host, secure_port, 1);
-            if (ld == NULL )
-                throw MakeStringException(-1, "ldap_sslinit error" );
-
-            int rc = 0;
-            unsigned long version = LDAP_VERSION3;
-            long lv = 0;
-            
-            rc = ldap_set_option(ld,
-                    LDAP_OPT_PROTOCOL_VERSION,
-                    (void*)&version);
-            if (rc != LDAP_SUCCESS)
-                throw MakeStringException(-1, "ldap_set_option error - %s", ldap_err2string(rc));
-
-            rc = ldap_get_option(ld,LDAP_OPT_SSL,(void*)&lv);
-            if (rc != LDAP_SUCCESS)
-                throw MakeStringException(-1, "ldap_get_option error - %s", ldap_err2string(rc));
-
-            // If SSL is not enabled, enable it.
-            if ((void*)lv != LDAP_OPT_ON)
-            {
-                rc = ldap_set_option(ld, LDAP_OPT_SSL, LDAP_OPT_ON);
-                if (rc != LDAP_SUCCESS)
-                    throw MakeStringException(-1, "ldap_set_option error - %s", ldap_err2string(rc));
-            }
-            
-            ldap_set_option(ld, LDAP_OPT_SERVER_CERTIFICATE, verifyServerCert);
+        bvalues = ldap_get_values_len(ld, msg, (const PCHAR)attr);
 #else
-            // Initialize an LDAP session for TLS/SSL
-            #ifndef HAVE_TLS
-                //throw MakeStringException(-1, "openldap client library libldap not compiled with TLS support");
-            #endif
-            StringBuffer uri("ldaps://");
-            uri.appendf("%s:%d", host, secure_port);
-            DBGLOG("connecting to %s", uri.str());
-            int rc = ldap_initialize(&ld, uri.str());
-            if(rc != LDAP_SUCCESS)
-            {
-                DBGLOG("ldap_initialize error %s", ldap_err2string(rc));
-                throw MakeStringException(-1, "ldap_initialize error %s", ldap_err2string(rc));
-            }
-            int reqcert = LDAP_OPT_X_TLS_NEVER;
-            ldap_set_option(NULL, LDAP_OPT_X_TLS_REQUIRE_CERT, &reqcert);
+        bvalues = ldap_get_values_len(ld, msg, attr);
 #endif
-        }
-        else
-        {
-            // Initialize an LDAP session
-            if ((ld = ldap_init( (char*)host, port )) == NULL)
-            {
-                throw MakeStringException(-1, "ldap_init error");
-            }
-        }
-        return ld;
-    }
-
-    static int LdapSimpleBind(LDAP* ld, char* userdn, char* password)
-    {
-#ifndef _WIN32
-        TIMEVAL timeout = {LDAPTIMEOUT, 0};
-        ldap_set_option(ld, LDAP_OPT_TIMEOUT, &timeout);
-        ldap_set_option(ld, LDAP_OPT_NETWORK_TIMEOUT, &timeout);
-#endif
-        return ldap_simple_bind_s(ld, userdn, password);
-        /*
-        //TODO: bugs need to be fixed: (1) in ldap_result, is "1" actually meant for LDAP_MESSAGE_ONE? (2) should call ldap_msgfree on result
-        int final_rc  = LDAP_SUCCESS;
-        int msgid = ldap_simple_bind(ld, userdn, password); 
-        if(msgid < 0)
-        {
-#ifndef _WIN32
-            final_rc = ldap_get_lderrno(ld, NULL, NULL);
-#else
-            final_rc = LDAP_OTHER;
-#endif
-        }
-        else
-        {
-            LDAPMessage* result = NULL;
-            TIMEVAL timeOut = {LDAPTIMEOUT,0};   
-            int rc = ldap_result(ld, msgid, 1, &timeOut, &result); 
-            if(rc < 0)
-            {
-#ifndef _WIN32
-                final_rc = ldap_get_lderrno(ld, NULL, NULL);
-#else
-                final_rc = LDAP_OTHER;
-#endif
-            }
-            else if(rc == 0)
-            {
-                final_rc = LDAP_TIMEOUT;
-            }
-            else
-            {
-                final_rc = ldap_result2error(ld, result, 1);
-            }
-        }
-        return final_rc;
-        */
-    }
-
-    // userdn is required for ldap_simple_bind_s, not really necessary for ldap_bind_s.
-    static int LdapBind(LDAP* ld, const char* domain, const char* username, const char* password, const char* userdn, LdapServerType server_type, const char* method="kerboros")
-    {
-        bool binddone = false;
-        int rc = LDAP_SUCCESS;
-        // By default, use kerberos authentication
-        if((method == NULL) || (strlen(method) == 0) || (stricmp(method, "kerberos") == 0))
-        {
-#ifdef _WIN32
-            if(server_type == ACTIVE_DIRECTORY)
-            {
-                if(username != NULL)
-                {
-                    SEC_WINNT_AUTH_IDENTITY secIdent;
-                    secIdent.User = (unsigned char*)username;
-                    secIdent.UserLength = strlen(username);
-                    secIdent.Password = (unsigned char*)password;
-                    secIdent.PasswordLength = strlen(password);
-                    // Somehow, setting the domain makes it slower
-                    secIdent.Domain = (unsigned char*)domain;
-                    secIdent.DomainLength = strlen(domain);
-                    secIdent.Flags = SEC_WINNT_AUTH_IDENTITY_ANSI;
-                    int rc = ldap_bind_s(ld, (char*)userdn, (char*)&secIdent, LDAP_AUTH_NEGOTIATE);
-                    if(rc != LDAP_SUCCESS)
-                    {
-                        DBGLOG("ldap_bind_s for user %s failed with %d - %s.", username, rc, ldap_err2string(rc));
-                        return rc;
-                    }
-                }
-                else
-                {
-                    int rc = ldap_bind_s(ld, NULL, NULL, LDAP_AUTH_NEGOTIATE);
-                    if(rc != LDAP_SUCCESS)
-                    {
-                        DBGLOG("User Authentication Failed - ldap_bind_s for current user failed with %d - %s.", rc, ldap_err2string(rc));
-                        return rc;
-                    }
-                }
-                binddone = true;
-            }
-#endif
-        }
-
-        if(!binddone)
-        {
-            if(userdn == NULL)
-            {
-                DBGLOG("userdn can't be NULL in order to bind to ldap server.");
-                return LDAP_INVALID_CREDENTIALS;
-            }
-            int rc = LdapSimpleBind(ld, (char*)userdn, (char*)password);
-            if (rc != LDAP_SUCCESS && server_type == OPEN_LDAP && strchr(userdn,','))
-            {   //Fedora389 is happier without the domain component specified
-                StringBuffer cn(userdn);
-                cn.replace(',',(char)NULL);
-                if (cn.length())//disallow call if no cn
-                    rc = LdapSimpleBind(ld, (char*)cn.str(), (char*)password);
-            }
-            if (rc != LDAP_SUCCESS )
-            {
-                // For Active Directory, try binding with NT format username
-                if(server_type == ACTIVE_DIRECTORY)
-                {
-                    StringBuffer logonname;
-                    logonname.append(domain).append("\\").append(username);
-                    rc = LdapSimpleBind(ld, (char*)logonname.str(), (char*)password);
-                    if(rc != LDAP_SUCCESS)
-                    {
-#ifdef LDAP_OPT_DIAGNOSTIC_MESSAGE
-                        char *msg=NULL;
-                        ldap_get_option(ld, LDAP_OPT_DIAGNOSTIC_MESSAGE, (void*)&msg);
-                        DBGLOG("LDAP bind error for user %s with %d - %s. %s", logonname.str(), rc, ldap_err2string(rc), msg&&*msg?msg:"");
-                        ldap_memfree(msg);
-#else
-                        DBGLOG("LDAP bind error for user %s with 0x%" I64F "x - %s", username, (unsigned __int64) rc, ldap_err2string(rc));
-#endif
-                        return rc;
-                    }
-                }
-                else
-                {
-                    DBGLOG("LDAP bind error for user %s with 0x%" I64F "x - %s", username, (unsigned __int64) rc, ldap_err2string(rc));
-                    return rc;
-                }
-            }
-        }
-        
-        return rc;
-    }
-
-    static void bin2str(MemoryBuffer& from, StringBuffer& to)
-    {
-        const char* frombuf = from.toByteArray();
-        char tmp[3];
-        for(unsigned i = 0; i < from.length(); i++)
-        {
-            unsigned char c = frombuf[i];
-            sprintf(tmp, "%02X", c);
-            tmp[2] = 0;
-            to.append("\\").append(tmp);
-        }
-    }
-
-    static int getServerInfo(const char* ldapserver, int ldapport, StringBuffer& domainDN, LdapServerType& stype, const char* domainname);
-
-    static void normalizeDn(const char* dn, const char* basedn, StringBuffer& dnbuf)
-    {
-        dnbuf.clear();
-        cleanupDn(dn, dnbuf);
-        if(!containsBasedn(dnbuf.str()))
-            dnbuf.append(",").append(basedn);
-    }
-
-    static bool containsBasedn(const char* str)
-    {
-        if(str == NULL || str[0] == '\0')
-            return false;
-        else
-            return (strstr(str, "dc=") != NULL);
-    }
-
-    static void cleanupDn(const char* dn, StringBuffer& dnbuf)
-    {
-        if(dn == NULL || dn[0] == '\0')
-            return;
-        dnbuf.append(dn);
-        dnbuf.toLowerCase();
-    }
-    
-    static bool getDcName(const char* domain, StringBuffer& dc)
-    {
-        bool ret = false;
-#ifdef _WIN32
-        PDOMAIN_CONTROLLER_INFO psInfo = NULL;
-        DWORD dwErr = DsGetDcName(NULL, domain, NULL, NULL, DS_FORCE_REDISCOVERY | DS_DIRECTORY_SERVICE_REQUIRED, &psInfo);
-        if( dwErr == NO_ERROR)
-        {
-            const char* dcname = psInfo->DomainControllerName;
-            if(dcname != NULL)
-            {
-                while(*dcname == '\\')
-                    dcname++;
-
-                dc.append(dcname);
-                ret = true;
-            }
-            NetApiBufferFree(psInfo);
-        }
-        else
-        {
-            DBGLOG("Error getting domain controller, error = %d", dwErr);
-            ret = false;
-        }
-#endif
-        return ret;
-    }
-
-    static void getName(const char* dn, StringBuffer& name)
-    {
-        const char* bptr = dn;
-        while(*bptr != '\0' && *bptr != '=')
-            bptr++;
-        
-        if(*bptr == '\0')
-        {
-            name.append(dn);
-            return;
-        }
-        else
-            bptr++;
-
-        const char* colon = strstr(bptr, ",");
-        if(colon == NULL)
-            name.append(bptr);
-        else
-            name.append(colon - bptr, bptr);
+        for (numValues = 0; bvalues && bvalues[numValues]; numValues++);
     }
 };
